Remvoing urls hack (#189)
The feature is being added upstream in 7.1 and it looks to be the cause of a significant number of major issues.

Falling back to the default urls will be done tonight 9-11 pm UTC.

Hopefully, this will set us up in a good place to get twitter working + then roll forward to py3.7.